feat(报价): 10.6.23.04

+++ business/src/main/java/com/xujie/business/domain/BO/BizEggQuotationDay.java
@@ -1,6 +1,5 @@
 package com.xujie.business.domain.BO;
 
-import java.math.BigDecimal;
 import java.util.Date;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -12,40 +11,30 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class BizEggQuotationDay {
-    private Long id;
+  private Long id;
 
-    private Long typeId;
+  private Long typeId;
 
-    /**
-    * 平均价
-    */
-    private BigDecimal avgNum;
+  /** 平均价 */
+  private Double avgNum;
 
-    /**
-    * 最高价
-    */
-    private BigDecimal maxNum;
+  /** 最高价 */
+  private Double maxNum;
 
-    /**
-    * 最低价
-    */
-    private BigDecimal minNum;
+  /** 最低价 */
+  private Double minNum;
 
-    /**
-    * 报价人数
-    */
-    private Integer peopleNumber;
+  /** 报价人数 */
+  private Integer peopleNumber;
 
-    /**
-    * 下降，持平，上升-1,0,1
-    */
-    private Integer status;
+  /** 下降，持平，上升-1,0,1 */
+  private Integer status;
 
-    private Date dataTime;
+  private Date dataTime;
 
-    private Date createTime;
+  private Date createTime;
 
-    private Date updateTime;
+  private Date updateTime;
 
-    private Integer isDelete;
+  private Integer isDelete;
 }
